Move utils to `/test/utils`
